feat(sql): refactor
