+++ housekeeper/cli/delete.py
@@ -7,8 +7,8 @@ from pathlib import Path
 import click
 
 from housekeeper.date import get_date
-from housekeeper.store.core import Store
 from housekeeper.store.models import File, Tag
+from housekeeper.store.store import Store
 
 LOG = logging.getLogger(__name__)
