+++ tests/cli/add/test_cli_add_file.py
@@ -7,8 +7,8 @@ from click import Context
 from click.testing import CliRunner
 
 from housekeeper.cli.add import file_cmd
-from housekeeper.store.core import Store
 from housekeeper.store.models import Bundle, Version
+from housekeeper.store.store import Store
 
 NEW_FILE_ADDED: str = "new file added"
