+++ tests/cli/add/test_cli_add_tags.py
@@ -3,8 +3,8 @@
 import logging
 
 from housekeeper.cli.add import tag_cmd
-from housekeeper.store.core import Store
 from housekeeper.store.models import File
+from housekeeper.store.store import Store
 
 
 def test_add_tags_no_args(populated_context, cli_runner, caplog):
