+++ tests/cli/delete/test_cli_delete_bundle.py
@@ -6,7 +6,7 @@ from click import Context
 from click.testing import CliRunner
 
 from housekeeper.cli import delete
-from housekeeper.store.core import Store
+from housekeeper.store.store import Store
 
 
 def test_delete_non_existing_bundle(base_context: Context, cli_runner: CliRunner, caplog):
