+++ tests/cli/delete/test_cli_delete_file.py
@@ -6,8 +6,8 @@ from click import Context
 from click.testing import CliRunner
 
 from housekeeper.cli import delete
-from housekeeper.store.core import Store
 from housekeeper.store.models import File
+from housekeeper.store.store import Store
 
 
 def test_delete_non_existing_file(base_context: Context, cli_runner: CliRunner, caplog):
