+++ tests/cli/delete/test_cli_delete_tag.py
@@ -6,8 +6,8 @@ from click import Context
 from click.testing import CliRunner
 
 from housekeeper.cli import delete
-from housekeeper.store.core import Store
 from housekeeper.store.models import File, Tag
+from housekeeper.store.store import Store
 
 
 def test_delete_non_existing_tag(
