+++ tests/cli/test_cli_include.py
@@ -8,8 +8,8 @@ from click.testing import CliRunner
 
 from housekeeper.cli.include import include
 from housekeeper.constants import ROOT
-from housekeeper.store.core import Store
 from housekeeper.store.models import Bundle, Version
+from housekeeper.store.store import Store
 
 
 def test_include_files_creates_bundle_dir(populated_context: dict, cli_runner: CliRunner):
