+++ tests/conftest.py
@@ -11,13 +11,13 @@ import pytest
 import yaml
 
 from housekeeper.date import get_date
-from housekeeper.store.core import Store
 from housekeeper.store.database import (
     create_all_tables,
     drop_all_tables,
     initialize_database,
 )
 from housekeeper.store.models import Archive, Bundle, Tag, Version
+from housekeeper.store.store import Store
 from tests.helper_functions import Helpers
 
 # basic fixtures
