+++ tests/store/crud/test_updatehandler.py
@@ -1,8 +1,8 @@
 from datetime import datetime
 
 from housekeeper import store
-from housekeeper.store.core import Store
 from housekeeper.store.models import Archive
+from housekeeper.store.store import Store
 from tests.helper_functions import Helpers
