+++ tests/store/filters/test_version_bundle_filters.py
@@ -2,11 +2,11 @@ import datetime as dt
 
 from sqlalchemy.orm import Query
 
-from housekeeper.store.core import Store
 from housekeeper.store.filters.version_bundle_filters import (
     filter_version_by_name_and_bundle_date,
 )
 from housekeeper.store.models import Version
+from housekeeper.store.store import Store
 
 
 def test_filter_version_by_name_and_bundle_date_returns_the_correct_version(populated_store: Store):
