+++ tests/store/filters/test_version_filters.py
@@ -1,8 +1,8 @@
 from sqlalchemy.orm import Query
 
-from housekeeper.store.core import Store
 from housekeeper.store.filters.version_filters import filter_version_by_id
 from housekeeper.store.models import Version
+from housekeeper.store.store import Store
 
 
 def test_get_version_query_returns_query_object(populated_store: Store):
